Adding hookRequest
